move streamrequest to models file, prewarm in webrtcd
