+++ system/athena/athenad.py
@@ -581,7 +581,7 @@ def getNetworks():
 
 @dispatcher.add_method
 def startStream(sdp: str, video_enabled: bool | None = None) -> dict:
-  from openpilot.system.webrtc.webrtcd import StreamRequestBody
+  from openpilot.system.webrtc.models import StreamRequestBody
   bridge_services_in = []
 
   # get live car params to avoid stale notCar edge case
